MRP-1440 tests: adjust test-framework to work with mpich2
Patch changes MACHINEFILE usage to be able to operate with mpich2

Reviewed-on: http://morpheus.xyus.xyratex.com:8443/gerrit/838
Tested-by: Jenkins

+++ lustre/tests/test-framework.sh
@@ -5135,6 +5135,9 @@ mixed_mdt_devs () {
 generate_machine_file() {
     local nodes=${1//,/ }
     local machinefile=$2
+
+    [[ -z $(echo $machinefile | tr -d " ") ]] && return 0
+
     rm -f $machinefile
     for node in $nodes; do
         echo $node >>$machinefile || \
